Fixes for TMA_Metrics-full.csv
Stop the inhibiting of TopdownL1, etc. metric writing.
:sup is used to count kernel cycles and should be the :k modifier for
perf. Handle the upper-case version of :SUP.
Allow metrics with the ##? operator by just removing as is done for the
 ## operator. Make this behavior more obvious by matching the full ## in
one go.
Drop the :percore modifier that Linux perf doesn't support.
Avoid an empty metric in the group with a ; at the start.