Add metrics from intel/perfmon-metrics
When reading the mapfile.csv, see if a perfmon-metrics directory exists.
If it does then add these metrics to the appropriate architecture. As
some of the metrics fail on upstream perf they are dropped, see the data
structure 'broken_extra_metrics'. If duplicate metrics exist then the
perfmon-metrics version is preferred.

Tested on BDX, CLX, ICX, SKX, SPR.