Extend the unsupported filter expressions
Filter expressions need to be in a format that works for perf or get
fixed via the csv file. When generating uncore events for architectures
like ivytown just CHAFilter was insufficient to generate an event list
that worked with perf. Add a notion of two kinds of filter, one for
events that should be dropped the other for the filter should just be
removed.

This is based on discussion with Kan Liang here:
https://lore.kernel.org/lkml/787d208d-accd-99f6-065b-ba1f914dcd25@linux.intel.com/