Add new metrics “UNCORE_FREQ” for Linux perf
In the Latest TMA file, the metric “Socket_CLKS" is the only metric for
the uncore frequency. It has already been supported in the Linux mainline
for a long time (since 2019)

 #perf stat -M "Socket_CLKS" -a sleep 1

Performance counter stats for 'system wide':

820,758,851 cha_0/event=0x0/ # 820758851.00 Socket_CLKS
1,001,926,584 ns duration_time

1.001926584 seconds time elapsed

The test result of Socket_CLKS is “820758851.00” (unit: HZ), the result is
not good for the user to get the direct information for uncore frequency.
Since the “Socket_CLKS” is already defined in the TMA, we should not change
its formula. We can create new metrics “UNCORE_FREQ” for Linux perf

UNCORE_FREQ = Socket_CLKS / #num_die/ duration_time / 1000000000 (GHz)

This new metric can be added to all platforms that have Sockets_CLKS.

Signed-off-by: Zhengjun Xing <zhengjun.xing@linux.intel.com>