Fix metrics error caused by "ScaleUnit"
The memory_bandwidth_read metrics report the wrong value.
If an event with “ScaleUnit”: “64Bytes”, the perf tool will automatically scale the raw value to the value of the byte by * 64.
But the memory_bandwidth_read metrics assume the raw value. It * 64 again.
{
        "BriefDescription": "DDR memory read bandwidth (MB/sec)",
        "MetricExpr": "( UNC_M_CAS_COUNT.RD * 64 / 1000000) / duration_time",
        "MetricGroup": "",
        "MetricName": "memory_bandwidth_read",
        "ScaleUnit": "1MB/s"
    },
The metrics are converted from the TMA data sheet, we need to keep it.
in perf-uncore-events-xxx.csv, we support both the original name and the new name for some events, For example, UNC_M_CAS_COUNT.RD is
the original name and "LLC_MISSES.MEM_READ" is the new name, we can keep UNC_M_CAS_COUNT.RD uses the raw data and
LLC_MISSES.MEM_READ applies the “64Bytes”, then its metrics can work ok.
Fix the same issue for similar uncore events.

Signed-off-by: Zhengjun Xing <zhengjun.xing@linux.intel.com>