Use #core_wide for #EBS_Mode
EBS_Mode is used in metrics like slots, to turn the clock domain from
core to hyperthread. The previous behavior is to generate two metrics
when #EBS_Mode is encountered, the _SMT suffixed version should be
used when hyperthreading is enabled. This approach is sub-optimal as
the hyperthread version corrects for what is running on the other
hyperthread even if running in system wide mode.

In the new approach a single metric is generated, #SMT_On is passed
through to perf and #EBS_Mode is converted to testing for #core_wide
being zero - indicating that the hyperthreads on a core aren't all
recording in system wide mode. For Core_Clks this gives 3
alternatives: #SMT_On is zero so core counters are valid; #EBS_Mode is
true, #core_wide is zero, so additional counters and adjustment are
necessary; #EBS_Mode is false, so the core counters are valid again
but need adjusting as each hyperthread will have a counter doubling
the expected clocks.