Generate all topdown levels
Remove the level == 1 test. Record a the parents of each topdown node so
that each child node can appear in a _group named after their parent.
Rename the metrics to have a tma_ prefix and to be lower case. Perf is
case insensitive wrt names and the tma_ prefix helps groups these
related metrics.