Remove expr-events
expr-events was an attempt to not expand large metrics, the value given
to it was the number counters before just using the original metric was
better. Unfortunately the implementation was buggy, allowing metrics to
refer to non-existent other metrics, and so the count was always set so
large that no use of metrics ever took place. Remove the implementation
as a later patch will implement the approach properly.